+++ packages/@runlightyear/gcal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/gcal
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [862932a7]
+  - @runlightyear/lightyear@0.29.0
+
 ## 0.5.2
 
 ### Patch Changes