+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/slack
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [862932a7]
+  - @runlightyear/lightyear@0.29.0
+
 ## 0.13.1
 
 ### Patch Changes