+++ packages/@runlightyear/smtp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/smtp
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [862932a7]
+  - @runlightyear/lightyear@0.29.0
+
 ## 0.1.15
 
 ### Patch Changes